Added a static method in the Statement class that checks whether a variable/ array name is a keyword. This is used in the DeclareStatement class and the DimStatement class.
Changed some code so that arrays and variables in the same code block cannot have the same name.
A variable can still be declared with a particular name if no array exists in the same code block with the same name, but an array with the same name exists somewhere outside that particular code block.