One can now redefine variables declared outside a subroutine block. Implemented that by changing the parentBlock of the subroutine code block to a new blank block with parentBlock set to null, and added all external variables and subroutine and function scopes there. That was easy :D.
variablesNotToFree is not really needed anymore for now. But I will leave it in just in case I want to add the ability to pass variables or arrays by reference. Right now you can only pass an expression. If you simply type the name of a variable its value is copied, therefore it is passed by value